#140 fixed the temporal workflow
