+++ knowledge_flow_backend/app/core/processors/input/docx_markdown_processor/tests/docx_markdown_processor_test.py
@@ -15,7 +15,7 @@
 import tempfile
 from pathlib import Path
 
-from app.common.structures import DocumentMetadata
+from app.common.document_structures import DocumentMetadata
 import pytest
 
 from app.core.processors.input.docx_markdown_processor.docx_markdown_processor import DocxMarkdownProcessor
