+++ knowledge_flow_backend/app/core/processors/output/empty_output_processor.py
@@ -16,7 +16,7 @@
 
 import logging
 
-from app.common.structures import DocumentMetadata
+from app.common.document_structures import DocumentMetadata
 from app.core.processors.output.base_output_processor import BaseOutputProcessor
 
 logger = logging.getLogger(__name__)
