+++ knowledge_flow_backend/app/core/processors/output/vectorization_processor/vectorization_processor.py
@@ -13,15 +13,15 @@
 # limitations under the License.
 
 import logging
+from app.common.structures import OutputProcessorResponse, Status
 from fastapi import HTTPException
 from langchain.schema.document import Document
 
 from app.application_context import ApplicationContext
-from app.common.structures import DocumentMetadata, Status, OutputProcessorResponse
+from app.common.document_structures import DocumentMetadata
 from app.core.processors.output.base_output_processor import BaseOutputProcessor
 
 logger = logging.getLogger(__name__)
-
 
 class VectorizationProcessor(BaseOutputProcessor):
     """
