+++ knowledge_flow_backend/app/core/stores/metadata/base_metadata_store.py
@@ -15,7 +15,7 @@
 from abc import abstractmethod
 from typing import List
 
-from app.common.structures import DocumentMetadata
+from app.common.document_structures import DocumentMetadata
 from app.core.stores.metadata.base_catalog_store import BaseCatalogStore
 
 class MetadataDeserializationError(Exception):
