+++ knowledge_flow_backend/app/core/stores/vector/base_vector_store.py
@@ -51,7 +51,7 @@ This design allows easy switching of backends (e.g., OpenSearch ➔ ChromaDB, Az
 
 from abc import ABC, abstractmethod
 from typing import List, Tuple
-from app.common.structures import DocumentMetadata
+from app.common.document_structures import DocumentMetadata
 from langchain.schema.document import Document
