+++ knowledge_flow_backend/app/features/catalog/controller.py
@@ -1,4 +1,4 @@
-from app.common.structures import DocumentSourceConfig, PullSourceConfig
+from app.common.structures import DocumentSourceConfig
 from fastapi import APIRouter, HTTPException, Query
 from typing import List, Literal, Optional
 from app.core.stores.metadata.base_catalog_store import PullFileEntry
